requirements.txt
Django>=4.2
djangorestframework
django-filter
django-crispy-forms
whitenoise
gunicorn
requests
python-dotenv
channels
channels-redis